Simplify RenderElement's shouldRepaintForImageAnimation() https://bugs.webkit.org/show_bug.cgi?id=140890
Reviewed by Darin Adler.

Simplify RenderElement's shouldRepaintForImageAnimation() by
not calling rendererForRootBackground(). The previous code was
doing a null-check on Document::documentElement() but failing
to do a null-check on the renderer just after, which lead me
to refactor this code.

Instead of calling document().documentElement.renderer(), we
now call renderer.parent(). This is equivalent because we already
know that renderer.isBody() returns true. The parent of the <body>
has to be the <html> element, which has to be the documentElement.
There is also no need to do a null-check on renderer.parent(). We
already know that the <body> has a renderer (because |renderer| is
its renderer), thus its parent <html> element has to have a renderer
as well.

Finally, simply call !rootElement.hasBackground() instead of
rootElement.rendererForRootBackground() == &bodyRenderer to determine
if the background painted by the root. This is equivalent but more
efficient because:
- We already know that the root renderer's is an <html> element. We
  can thus avoid the check in rendererForRootBackground().
- We already have the <body>'s renderer so we don't need to do the
  DOM tree traversal in rendererForRootBackground() to get it.

No new tests, already covered by:
fast/images/animated-gif-body-delegated-background-image.html
fast/images/animated-gif-html-background-image.html

* rendering/RenderElement.cpp:
(WebCore::shouldRepaintForImageAnimation):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179145 268f45cc-cd09-0410-ab3c-d52691b4dbfc