Parser::parseVarDeclarationList gets the wrong JSToken for the last identifier https://bugs.webkit.org/show_bug.cgi?id=141272
Reviewed by Oliver Hunt.

This patch fixes a bug where the wrong text location would be 
assigned to a variable declaration inside a ForIn/ForOf loop. 
It also fixes a bug in the type profiler where the type profiler 
emits the wrong text offset for a ForIn loop's variable declarator 
when it's not a pattern node.

* bytecompiler/NodesCodegen.cpp:
(JSC::ForInNode::emitLoopHeader):
* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseVarDeclarationList):
* tests/typeProfiler/loop.js:
(testForIn):
(testForOf):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179873 268f45cc-cd09-0410-ab3c-d52691b4dbfc