Layout Test js/regress-141098.html is failing on 32-bit Machines https://bugs.webkit.org/show_bug.cgi?id=141848
Reviewed by Geoffrey Garen.

It appears that different control flow paths in probeAndRecurse() allowed the second time
through the test to recurse deeper before getting to the point of overflowing the stack.
Restructured the test so that the exact same control flow in probeAndRecurse() is used
both times we call it, including probing the depth of the stack.  Now we pass a flag that
indicates whether or not we should try ever expanding eval strings or reuse the most
recent eval string.

* js/script-tests/regress-141098.js:
(testEval):
(probeAndRecurse):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@180453 268f45cc-cd09-0410-ab3c-d52691b4dbfc