<attachment> should support indication of download progress https://bugs.webkit.org/show_bug.cgi?id=142336 <rdar://problem/19982504>
Reviewed by Anders Carlsson.

* html/HTMLAttachmentElement.cpp:
(WebCore::HTMLAttachmentElement::parseAttribute):
Invalidate the attachment if the progress attribute changes.

* platform/graphics/FloatRoundedRect.h:
(WebCore::FloatRoundedRect::Radii::Radii):
Add a Radii constructor that takes a single argument that is used
as the radius for all corners, for convenience.

* rendering/RenderThemeMac.mm:
(WebCore::attachmentProgressBarBackgroundColor):
(WebCore::attachmentProgressBarFillColor):
(WebCore::attachmentProgressBarBorderColor):
Add a bunch of constants for the progress bar.

(WebCore::paintAttachmentIconBackground):
Make use of fillRoundedRect instead of creating a path.

(WebCore::paintAttachmentProgress):
(WebCore::RenderThemeMac::paintAttachment):
Paint a progress bar if the progress attribute exists and
is a valid floating point number.

* fast/attachment/attachment-progress.html: Added.
* platform/mac/fast/attachment/attachment-progress-expected.png: Added.
* platform/mac/fast/attachment/attachment-progress-expected.txt: Added.
* platform/mac-mavericks/fast/attachment/attachment-progress-expected.txt: Added.
Add a test for the <attachment> progress attribute.

* platform/mac/fast/attachment/attachment-rendering-expected.png:
Rebaseline a test that changed in r181058 (and will change again!), but
which didn't start failing because the pixel test tolerance is such that
it only fails the hash check and not the test itself.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@181114 268f45cc-cd09-0410-ab3c-d52691b4dbfc