Web Inspector: Better handle displaying -0 https://bugs.webkit.org/show_bug.cgi?id=142708
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-03-15
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

Modeled after a blink change:

Patch by <aandrey@chromium.org>
DevTools: DevTools: Show -0 for negative zero in console
https://src.chromium.org/viewvc/blink?revision=162605&view=revision

* inspector/InjectedScriptSource.js:
When creating a description string, or preview value string
for -0, be sure the string is "-0" and not "0".

LayoutTests:

Add tests for -0 in different places.
I don't understand yet why the test says "0" for
the property previews of "-0". Everything behaves
correctly, but I can't see to make the test show
the right value appear in the test. That is worth
investigating separately though.

* inspector/model/remote-object-expected.txt:
* inspector/model/remote-object.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@181516 268f45cc-cd09-0410-ab3c-d52691b4dbfc