Source/WebCore: WebCore part of <rdar://problem/20282863> Transforms are flattened in snapshots of on-screen WKWebViews https://bugs.webkit.org/show_bug.cgi?id=143023
Reviewed by Tim Horton.

* platform/spi/cocoa/QuartzCoreSPI.h: Added the declaration of
CARenderServerRenderLayerWithTransform.

Source/WebKit2:
WebKit2 part of <rdar://problem/20282863> Transforms are flattened in snapshots of on-screen WKWebViews
https://bugs.webkit.org/show_bug.cgi?id=143023

Reviewed by Tim Horton.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _snapshotRect:intoImageOfWidth:completionHandler:]): If the view is in a
window, use CARenderServerRenderLayerWithTransform to synchronously capture a snapshot of
its layer tree into an IOSurfcae, then call the completion handler with an image created
from the IOSurface.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@181910 268f45cc-cd09-0410-ab3c-d52691b4dbfc