REGRESSION(r180000): Changing moving average or enveloping strategy doesn't update the graph https://bugs.webkit.org/show_bug.cgi?id=143254
Reviewed by Antti Koivisto.

The bug was caused by App.Pane no longer replacing 'chartData' property when updating the moving average
or the enveloping values. Fixed the bug by creating a new chartData object when the strategy is changed
so that the interactive chart component will observe a change to 'chartData'.

* public/v2/app.js:
(App.Pane._movingAverageOrEnvelopeStrategyDidChange): Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182184 268f45cc-cd09-0410-ab3c-d52691b4dbfc