Add Options::jitPolicyScale() as a single knob to make all compilations happen sooner. <https://webkit.org/b/143385>
Reviewed by Geoffrey Garen.

For debugging purposes, sometimes, we want to be able to make compilation happen
sooner to see if we can accelerate the manifestation of certain events / bugs.
Currently, in order to achieve this, we'll have to tweak multiple JIT thresholds
which make up the compilation policy.  Let's add a single knob that can tune all
the thresholds up / down in one go proportionately so that we can easily tweak
how soon compilation occurs.

* runtime/Options.cpp:
(JSC::scaleJITPolicy):
(JSC::recomputeDependentOptions):
* runtime/Options.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182337 268f45cc-cd09-0410-ab3c-d52691b4dbfc