Rename ActiveDOMObject::canSuspend() to canSuspendForPageCache() for clarity https://bugs.webkit.org/show_bug.cgi?id=143513
Reviewed by Andreas Kling.

Source/WebCore:

Rename ActiveDOMObject::canSuspend() to canSuspendForPageCache() for
clarity as it is only used for the page cache. Also rename the
'DocumentWillBecomeInactive' value of the ReasonForSuspension enum to
'PageCache' as it is only used by CachedFrame for the page cache and it
is a lot more understandable.

Source/WebKit/mac:

Update ScriptExecutionContext::canSuspendActiveDOMObjects() to
canSuspendActiveDOMObjectsForPageCache() as the method was renamed.

* WebView/WebFrame.mm:
(-[WebFrame _cacheabilityDictionary]):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182544 268f45cc-cd09-0410-ab3c-d52691b4dbfc