Rename JSC_dfgFunctionWhitelistFile to JSC_dfgWhitelist. https://bugs.webkit.org/show_bug.cgi?id=143722
Reviewed by Michael Saboff.

Renaming JSC_dfgFunctionWhitelistFile to JSC_dfgWhitelist so that it is
shorter, and easier to remember (without having to look it up) and to
type.  JSC options now support descriptions, and one can always look up
the description if the option's purpose is not already obvious.

* dfg/DFGFunctionWhitelist.cpp:
(JSC::DFG::FunctionWhitelist::ensureGlobalWhitelist):
(JSC::DFG::FunctionWhitelist::contains):
* runtime/Options.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182811 268f45cc-cd09-0410-ab3c-d52691b4dbfc